MNT: move size_from_uri from tiled.utils to tiled.storage
Per review feedback on #570: avoid growing the catch-all tiled.utils
module. size_from_uri is intrinsically about storage backends (it
dispatches on URI scheme to stat local files or HEAD object-store blobs),
so it belongs in tiled.storage alongside the Storage classes and
get_storage(). The move also drops the deferred-import workaround that
existed because tiled.storage imports from tiled.utils.

Tests moved to tests/test_storage.py.
